Get libraries for matching with specific mode and type

Get libraries for matching with specific mode and type

get_libs(mode = c("Pos", "Neg"), acq = c("dda", "aif"))

Arguments

mode
Acquisition mode. Default is 'Pos'

acq
Acquisition type, either dda or aif. Default is 'dda'

Value

A tibble containing all libraries for the chosen mode and type.

Examples

get_libs(mode = 'Pos', acq = 'dda')

#> # A tibble: 83 x 12
#>    file  and_cols or_cols dda   aif   class_only mode  adduct class_name ions 
#>    <chr> <list>   <list>  <lgl> <lgl> <lgl>      <chr> <chr>  <chr>      <nam>
#>  1 AcCa… <chr [2… <chr [… TRUE  TRUE  FALSE      Pos   [M+H]+ Acylcarni… <tib…
#>  2 CE_N… <chr [2… <chr [… TRUE  TRUE  FALSE      Pos   [M+NH… Cholester… <tib…
#>  3 CerP… <chr [2… <chr [… TRUE  FALSE FALSE      Pos   [M+H]+ Ceramide-… <tib…
#>  4 CL_3… <chr [4… <chr [… TRUE  TRUE  FALSE      Pos   [M+3N… Cardiolip… <tib…
#>  5 CL_N… <chr [2… <chr [… TRUE  TRUE  FALSE      Pos   [M+NH… Cardiolip… <tib…
#>  6 CoQ_… <chr [1… <chr [… TRUE  TRUE  FALSE      Pos   [M+NH… Coenzyme Q <tib…
#>  7 DG_N… <chr [2… <chr [… TRUE  TRUE  FALSE      Pos   [M+NH… Diacylgly… <tib…
#>  8 DGDG… <chr [1… <chr [… TRUE  FALSE FALSE      Pos   [M+NH… Digalacto… <tib…
#>  9 DGDG… <chr [1… <chr [… TRUE  FALSE TRUE       Pos   [M+NH… Digalacto… <tib…
#> 10 DGTS… <chr [3… <chr [… TRUE  FALSE FALSE      Pos   [M+H]+ Diacylgly… <tib…
#> # … with 73 more rows, and 2 more variables: user_defined <lgl>, classkmd <dbl>