Optimize matching lipids using Referenced Kendrick Mass Defects

Optimize matching lipids using Referenced Kendrick Mass Defects

match_kmd(results, kmd_cutoff = 0.2)

Arguments

results
Annotated features obtoined from [match_ms1], [match_ms2] or [merge_ms2].

kmd_cutoff
Maximum tolerated error between the number of double bonds in a matched lipid and the expected number, as caculated from Referenced Kendrick Mass Defect.

Value

Annotated features with best_match updated to only molecules that pass the kmd_cutoff criteria

Examples

## With MS2 data only
ms2_file <- system.file("extdata", "ms2file.ms2", package = "lipID")

ms2_data <- read_ms2(ms2_file)
libs <- get_libs()
confirmed_molecules <- match_ms2(ms2_data, libs)

#> Joining, by = c("mz", "ms2_rt")

#> Joining, by = c("precursor", "ms2_rt")

match_kmd(confirmed_molecules, kmd_cutoff = 0.2)

#> Joining, by = "file"

#> # A tibble: 2,970 x 24
#>    ms2_file precursor ms2_rt name  confirmed ions_matched fragments_inten…
#>    <fct>        <dbl>  <dbl> <chr> <lgl>     <chr>                   <dbl>
#>  1 ms2file       665.   28.6 MGDG… FALSE     [M+NH4-CO]+           688019.
#>  2 ms2file       665.   28.5 MGDG… FALSE     [M+NH4-CO]+           385381 
#>  3 ms2file       691.   30.7 MGDG… FALSE     [M+NH4-CO]+           100997.
#>  4 ms2file       771.   35.1 MGDG… FALSE     [M+NH4-CO]+            83109.
#>  5 ms2file       691.   30.6 MGDG… FALSE     sn1_NL;sn2_…            3358.
#>  6 ms2file       837.   29.8 SQDG… FALSE     [M-C6H10O7S…           25054.
#>  7 ms2file       837.   29.8 SQDG… FALSE     [M-C6H10O7S…           24287.
#>  8 ms2file       839.   30.4 SQDG… FALSE     [M-C6H10O7S…            6153.
#>  9 ms2file       791.   31.0 Plas… FALSE     PC_Head                 5094.
#> 10 ms2file       791.   31.0 Plas… FALSE     PC_Head                 5094.
#> # … with 2,960 more rows, and 17 more variables: partial_match <dbl>,
#> #   file <chr>, class_name <chr>, sum_composition <chr>, odd_chain <lgl>,
#> #   modifs <chr>, total_cl <dbl>, total_cs <dbl>, n_and <int>, n_or <int>,
#> #   n_and_true <int>, n_or_true <int>, and_cols <lgl>, or_cols <lgl>,
#> #   best_match <lgl>, cs_residual <dbl>, cs_matched <lgl>

## With MS1 features
features_file <- system.file("extdata", "features.csv", package = "lipID")
features <- readr::read_csv(features_file)

#> Parsed with column specification:
#> cols(
#>   mz = col_double(),
#>   rt = col_double(),
#>   Sample1 = col_double(),
#>   Sample2 = col_double(),
#>   Sample3 = col_double(),
#>   Sample4 = col_double(),
#>   Sample5 = col_double(),
#>   Sample6 = col_double()
#> )

annotated_features <- merge_ms2(features, confirmed_molecules)

#> Joining, by = c("mz", "rt")

#> Joining, by = c("precursor", "ms2_rt")

match_kmd(annotated_features, kmd_cutoff = 0.2)

#> Joining, by = "file"

#> # A tibble: 2,239 x 33
#>       mz    rt name  ms2_file precursor ms2_rt best_match partial_match
#>    <dbl> <dbl> <chr> <fct>        <dbl>  <dbl> <lgl>              <dbl>
#>  1  747.  33.6 Plas… ms2file       747.   33.5 TRUE                   1
#>  2  747.  33.6 Plas… ms2file       747.   33.5 FALSE                  1
#>  3  783.  30.7 PC(1… ms2file       783.   31.1 TRUE                   1
#>  4  783.  30.3 PC(1… ms2file       783.   31.1 TRUE                   1
#>  5  811.  32.7 PC(1… ms2file       811.   33.2 TRUE                   1
#>  6  785.  31.9 PC(1… ms2file       785.   31.7 TRUE                   1
#>  7  785.  31.6 PC(1… ms2file       785.   31.7 TRUE                   1
#>  8  835.  32.1 PC(1… ms2file       835.   32.2 TRUE                   1
#>  9  745.  33.0 PE(1… ms2file       745.   33.0 TRUE                   1
#> 10  807.  31.6 PC(1… ms2file       807.   30.9 TRUE                   1
#> # … with 2,229 more rows, and 25 more variables: confirmed <lgl>,
#> #   ions_matched <chr>, fragments_intensity <dbl>, Sample1 <dbl>,
#> #   Sample2 <dbl>, Sample3 <dbl>, Sample4 <dbl>, Sample5 <dbl>, Sample6 <dbl>,
#> #   nearest_ms2 <lgl>, file <chr>, class_name <chr>, sum_composition <chr>,
#> #   odd_chain <lgl>, modifs <chr>, total_cl <dbl>, total_cs <dbl>, n_and <int>,
#> #   n_or <int>, n_and_true <int>, n_or_true <int>, and_cols <lgl>,
#> #   or_cols <lgl>, cs_residual <dbl>, cs_matched <lgl>